refactor: separated api and svc layer for events
