+++ internal/events/commons.go
@@ -1,4 +1,4 @@
-package commons
+package events
 
 type Header string
